requirements.txt
Django==5.0.3
djangorestframework==3.15.1
pandas==1.5.2